refactor: extract sync files domain
