+++ src/sync/domains/index.ts
@@ -1,5 +1,6 @@
 export * from "@/sync/domains/calendar";
 export * from "@/sync/domains/contacts";
+export * from "@/sync/domains/files";
 export * from "@/sync/domains/songs";
 export * from "@/sync/domains/stickies";
 export * from "@/sync/domains/videos";
